Added methoods to get Trsansactions & changed trensaction amount column to store debited/credited amount instead of balance

+++ Sweep/src/main/java/com/banking/Sweep/service/impl/UserServiceImpl.java
@@ -1,13 +1,14 @@
 package com.banking.Sweep.service.impl;
 
 import com.banking.Sweep.DTO.UserDTO;
+import com.banking.Sweep.Exception.DoesNotExistException;
+import com.banking.Sweep.Exception.DuplicateEmailException;
 import com.banking.Sweep.model.User;
 import com.banking.Sweep.repository.UserRepository;
 import com.banking.Sweep.service.UserService;
 import org.modelmapper.ModelMapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-
 import java.util.List;
 import java.util.Map;
 import java.util.stream.Collectors;
@@ -27,19 +28,19 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User createUser(User user) {
+    public void createUser(User user) {
+        if(userRepository.existsByUserEmail(user.getUserEmail())){
+            throw new DuplicateEmailException("The email address " + user.getUserEmail() + " is already in use.");
+        }
         userRepository.save(user);
-
-        return user;
     }
 
     @Override
     public UserDTO retriveUserById(Long userId) {
 
-        User user=userRepository.findById(userId).get();
+            User user = userRepository.findById(userId).orElseThrow(()->new DoesNotExistException("User not found with ID: " + userId));
+            return modelMapper.map(user, UserDTO.class);
 
-        return modelMapper.map(user,UserDTO.class);
-        //return userRepository.findByDTOId(userId);
     }
 
     @Override
@@ -53,13 +54,16 @@ public class UserServiceImpl implements UserService {
     @Override
     public void updateUser(Long userId, Map<String,String> updates) {
         User user = userRepository.findById(userId)
-                .orElseThrow(() -> new RuntimeException("User not found with ID: " + userId));
+                .orElseThrow(() -> new DoesNotExistException("User not found with ID: " + userId));
 
         // Update fields dynamically based on provided keys
         if (updates.containsKey("userName")) {
             user.setUserName(updates.get("userName"));
         }
         if (updates.containsKey("userEmail")) {
+            if(userRepository.existsByUserEmail(updates.get("userEmail")) && !user.getUserEmail().equals(updates.get("userEmail"))){
+                throw new DuplicateEmailException("The email address " + updates.get("userEmail") + " is already in use.");
+            }
             user.setUserEmail(updates.get("userEmail"));
         }
         userRepository.save(user);
@@ -67,12 +71,23 @@ public class UserServiceImpl implements UserService {
 
     @Override
     public void deleteUser(Long userId) {
-        userRepository.deleteById(userId);
+        if(userRepository.existsById(userId)){
+            userRepository.deleteById(userId);
+        }
+        else{
+            throw new DoesNotExistException("User not found with ID: " +userId+" to delete");
+        }
+
     }
 
     @Override
     public UserDTO getUserByEmail(String userEmail) {
-        return userRepository.findByEmail(userEmail);
+        if(userRepository.existsByUserEmail(userEmail)) {
+            return userRepository.findByEmail(userEmail);
+        }
+        else{
+            throw new DoesNotExistException("User not found with email :"+userEmail);
+        }
     }
